refactor(lib): move WatchedPath out of fs
